requirements.txt
asgiref==3.7.2
dj-database-url==2.1.0
dj-email-url==1.0.6
Django==5.0
django-cache-url==3.4.5
django-ckeditor==6.7.0
django-jet-reboot==1.3.5
django-js-asset==2.2.0
environs==10.0.0
gunicorn==21.2.0
marshmallow==3.20.1
packaging==23.2
pillow == 10.1.0
psycopg==3.1.16
psycopg-binary==3.1.16
python-dotenv==1.0.0
sqlparse==0.4.4
typing_extensions==4.9.0
whitenoise==6.6.0
